+++ sockeye/inference_pt.py
@@ -31,7 +31,7 @@ from . import lexicon
 from . import utils
 from . import vocab
 from .beam_search_pt import CandidateScorer, get_search_algorithm, GreedySearch, SearchResult
-from .data_io_pt import tokens2ids
+from .data_io import tokens2ids
 from .model_pt import PyTorchSockeyeModel
 
 logger = logging.getLogger(__name__)
