+++ sockeye/prepare_data_pt.py
@@ -17,7 +17,7 @@ import os
 
 from . import arguments
 from . import constants as C
-from . import data_io_pt
+from . import data_io
 from . import utils
 from . import vocab
 from .log import setup_main_logger
