+++ sockeye/train_pt.py
@@ -33,7 +33,7 @@ import torch.distributed.elastic.multiprocessing.errors
 from . import arguments
 from . import checkpoint_decoder
 from . import constants as C
-from . import data_io_pt
+from . import data_io
 from . import encoder_pt
 from . import layers_pt
 from . import loss_pt
